address review: use pendingPVCError type and consolidate the test
- Introduce a pendingPVCError type (mirroring pendingDRAResourceClaimError)
  so the addPod retry path is scoped to the PVC-not-yet-synced case only,
  instead of a generic errors.IsNotFound check that could catch unrelated
  NotFound errors from other resource lookups in NewTaskInfo.
- Move the regression test into event_handlers_test.go and drop the
  standalone event_handlers_pvc_race_test.go file.
- Rewrite the test to assert the pod becomes a proper TaskInfo in the
  cache (carrying its pod reference) and is queued for resync, rather
  than exercising the resync-drain loop mechanics.
