+++ pkg/scheduler/cache/event_handlers.go
@@ -126,15 +126,14 @@ func (sc *SchedulerCache) getPodCSIVolumes(pod *v1.Pod) (map[v1.ResourceName]int
 
 		pvc, err := sc.pvcInformer.Lister().PersistentVolumeClaims(pod.Namespace).Get(pvcName)
 		if err != nil {
-			// The PVC is required to proceed with
-			// scheduling of a new pod because it cannot
-			// run without it. Bail out immediately.
-			//
-			// Wrap with %w so callers can distinguish
-			// "PVC not yet in informer" (not-found) from
-			// other lookup failures via errors.Is /
-			// apierrors.IsNotFound and recover accordingly.
-			return volumes, fmt.Errorf("looking up PVC %s/%s: %w", pod.Namespace, pvcName, err)
+			// The PVC is required to proceed with scheduling of a new
+			// pod because it cannot run without it. A PVC ADD event can
+			// reach the scheduler's PVC informer after the pod's ADD
+			// reaches the pod informer (e.g. a pod created together with
+			// an OnDemand PVC). addPod handles this by adding the task to
+			// the cache and re-queueing it for resync instead of dropping
+			// the pod.
+			return volumes, fmt.Errorf("looking up PVC %s/%s: %v", pod.Namespace, pvcName, err)
 		}
 		// The PVC for an ephemeral volume must be owned by the pod.
 		if isEphemeral {
@@ -276,19 +275,19 @@ func (sc *SchedulerCache) addPod(pod *v1.Pod) error {
 			sc.resyncTask(pi)
 			return nil
 		}
-		// Recover from a Pod ADD that races ahead of its PVC ADD on
-		// the informers. Same pattern as the DRA branch above.
-		if errors.IsNotFound(err) {
-			klog.V(4).Infof("PVC for pod <%s/%s> not yet in informer cache, add task and retry: %v", pod.Namespace, pod.Name, err)
-			if addErr := sc.addTask(pi); addErr != nil {
-				return addErr
-			}
-			sc.resyncTask(pi)
-			return nil
+		// NewTaskInfo can fail transiently, e.g. a referenced PVC has not
+		// yet reached the scheduler's informer when the pod ADD is
+		// processed. Always add the task to the cache and enqueue it for
+		// resync so syncTask re-fetches the pod and rebuilds the task once
+		// the dependency syncs, rather than dropping the pod: resyncTask on
+		// a task that was never added to sc.Jobs would be Forgotten by
+		// processResyncTask and lost.
+		klog.V(4).Infof("generate taskInfo for pod <%s/%s> failed, add task to cache and retry: %v", pod.Namespace, pod.Name, err)
+		if addErr := sc.addTask(pi); addErr != nil {
+			return addErr
 		}
-		klog.Errorf("generate taskInfo for pod(%s) failed: %v", pod.Name, err)
 		sc.resyncTask(pi)
-		return err
+		return nil
 	}
 
 	return sc.addTask(pi)
